Improve timing logic
Fix what amounts to an off-by-one error when calculating the average click interval from two or more clicks.

Change waiting for second click string to be ... to show user that we are waiting for their second click.

Upshot is, now it actually gives a roughly correct answer from the first two taps.